Wow, they changed the endpoint under our feet!
I wanted to do a USB trace on macOS for comparison's sake.
RØDE Central forced me to update the device (from firmware 2.1.9
to 2.2.0 iirc) before I could use it, and now the device uses a
different endpoint (0x83 instead of 0x82) for the interrupt
communication.

LOL, we'll need to redo the Windows traces as well, then!